% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{Defines logic for updating the app based on user interaction in the ui}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{see shiny app architecture}

\item{output}{see shiny app architecture}

\item{session}{see shiny app architecture}
}
\value{
server function for use in a shiny app
}
\description{
Defines logic for updating the app based on user interaction in the ui
}
