% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{boxplot_time_unit}
\alias{boxplot_time_unit}
\title{Generates a boxplot to visualize the distribution of time between
GPS measurements by GPS unit.}
\usage{
boxplot_time_unit(rds_path)
}
\arguments{
\item{rds_path}{Path of .rds animal data file to read in}
}
\value{
distribution of time between GPS measurements by GPS unit, as a boxplot
}
\description{
Generates a boxplot to visualize the distribution of time between
GPS measurements by GPS unit.
}
\examples{
# Boxplot of GPS measurement time differences for demo data .rds

boxplot_time_unit(system.file("extdata", "demo_aug19.rds", package = "animaltracker"))
}
