% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{calc_bearing}
\alias{calc_bearing}
\title{Helper function for cleaning Columbus P-1 datasets.
Given lat and long coords in degree decimal, convert to radians and compute bearing.}
\usage{
calc_bearing(lat1, lon1, lat2, lon2)
}
\arguments{
\item{lat1}{latitude of starting point}

\item{lon1}{longitude of starting point}

\item{lat2}{latitude of ending point}

\item{lon2}{longitude of ending point}
}
\value{
bearing computed from given coordinates
}
\description{
Helper function for cleaning Columbus P-1 datasets.
Given lat and long coords in degree decimal, convert to radians and compute bearing.
}
