% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_storage.R
\name{clean_batch}
\alias{clean_batch}
\title{Cleans a directory of animal data files and stores them in .rds files}
\usage{
clean_batch(data_dir)
}
\arguments{
\item{data_dir}{location of animal data files, in list format}
}
\value{
df of metadata for animal file directory
}
\description{
Cleans a directory of animal data files and stores them in .rds files
}
