% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_storage.R
\name{clean_df}
\alias{clean_df}
\title{Clean animal data frame}
\usage{
clean_df(df, ani_id, gps_id)
}
\arguments{
\item{df}{raw input data frame}

\item{ani_id}{animal ID (from meta)}

\item{gps_id}{GPS ID (from meta)}
}
\value{
cleaned data frame
}
\description{
Clean animal data frame
}
