% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{compare_flags}
\alias{compare_flags}
\title{Joins and reformats two animal data frames for the purpose of flag comparison}
\usage{
compare_flags(correct, candidate, elev = TRUE, slope = TRUE)
}
\arguments{
\item{correct}{reference df}

\item{candidate}{df to be compared to the reference}

\item{elev}{logical, whether to include elevation, defaults to true}

\item{slope}{logical, whether to include slope, defaults to true}
}
\value{
joined and reformatted df
}
\description{
Joins and reformats two animal data frames for the purpose of flag comparison
}
\examples{
# Join and reformat unfiltered demo data and filtered demo data

compare_flags(demo_unfiltered_elev, demo_filtered_elev)
}
