% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs_module.R
\name{datePicker}
\alias{datePicker}
\title{'shiny' module server-side UI generator for the animaltracker app's date picker.}
\usage{
datePicker(input, output, session, req_list, text)
}
\arguments{
\item{input}{'shiny' server input, automatically populated}

\item{output}{'shiny' server output, automatically populated}

\item{session}{'shiny' server session, automatically populated}

\item{req_list}{list of reactive statements required to display picker}

\item{text}{title for picker}
}
\value{
'shiny' renderUI object for date picker
}
\description{
'shiny' module server-side UI generator for the animaltracker app's date picker.
}
