% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs_module.R
\name{datePickerOutput}
\alias{datePickerOutput}
\title{'shiny' module UI output for the animaltracker app's date picker.}
\usage{
datePickerOutput(id)
}
\arguments{
\item{id}{chosen ID of UI output}
}
\value{
'shiny' uiOutput for date picker
}
\description{
'shiny' module UI output for the animaltracker app's date picker.
}
