% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs_module.R
\name{dateSlider}
\alias{dateSlider}
\title{Shiny Module server-side UI generator for the animaltracker app's date slider.}
\usage{
dateSlider(input, output, session, req_list, text)
}
\arguments{
\item{input}{Shiny server input, automatically populated}

\item{output}{Shiny server output, automatically populated}

\item{session}{Shiny server session, automatically populated}

\item{req_list}{list of reactive statements required to display slider}

\item{text}{title for slider}
}
\description{
Shiny Module server-side UI generator for the animaltracker app's date slider.
}
