% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{deg_to_dec}
\alias{deg_to_dec}
\title{Helper function for cleaning Columbus P-1 datasets.
Given lat or long coords in degrees and a direction, convert to decimal.}
\usage{
deg_to_dec(x, direction)
}
\arguments{
\item{x}{lat or long coords in degrees}

\item{direction}{direction of lat/long}
}
\value{
converted x
}
\description{
Helper function for cleaning Columbus P-1 datasets.
Given lat or long coords in degrees and a direction, convert to decimal.
}
