% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_comparison}
\alias{demo_comparison}
\title{Demo comparison of two animal datasets}
\format{
A data frame with 2758 rows and 33 variables
}
\usage{
demo_comparison
}
\description{
Demo comparison of two animal datasets
}
\keyword{datasets}
