% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_filtered}
\alias{demo_filtered}
\title{Filtered demo animal GPS data from cows}
\format{A data frame with 323278 rows and 26 variables}
\usage{
demo_filtered
}
\description{
Filtered demo animal GPS data from cows
}
\keyword{datasets}
