% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_filtered_elev}
\alias{demo_filtered_elev}
\title{Filtered demo animal GPS data from cows with elevation appended at zoom 1}
\format{A data frame with 2187 rows and 29 variables}
\usage{
demo_filtered_elev
}
\description{
Filtered demo animal GPS data from cows with elevation appended at zoom 1
}
\keyword{datasets}
