% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_info}
\alias{demo_info}
\title{Raw demo animal GPS data from cows with information}
\format{
A list with 10 elements
}
\usage{
demo_info
}
\description{
Raw demo animal GPS data from cows with information
}
\keyword{datasets}
