% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_meta}
\alias{demo_meta}
\title{Metadata for demo animal GPS data from cows}
\format{A data frame with 39 rows and 11 variables}
\usage{
demo_meta
}
\description{
Metadata for demo animal GPS data from cows
}
\keyword{datasets}
