% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_raw}
\alias{demo_raw}
\title{Raw demo animal GPS data from cows with information}
\format{A 39 MB list with 10 elements}
\usage{
demo_raw
}
\description{
Raw demo animal GPS data from cows with information
}
\keyword{datasets}
