% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_unfiltered}
\alias{demo_unfiltered}
\title{Unfiltered demo animal GPS data from cows}
\format{A data frame with 2288 rows and 32 variables}
\usage{
demo_unfiltered
}
\description{
Unfiltered demo animal GPS data from cows
}
\keyword{datasets}
