% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_unfiltered_elev}
\alias{demo_unfiltered_elev}
\title{Unfiltered demo animal GPS data from cows with elevation}
\format{A data frame with 2288 rows and 35 variables}
\usage{
demo_unfiltered_elev
}
\description{
Unfiltered demo animal GPS data from cows with elevation
}
\keyword{datasets}
