% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{detect_peak_knn}
\alias{detect_peak_knn}
\title{Train a k-nearest neighbors model on a comparison data frame to classify points that should be dropped}
\usage{
detect_peak_knn(df_comparison, normalize_type = "standardize",
  train = 2/3)
}
\arguments{
\item{df_comparison}{output of compare_flags}

\item{normalize_type}{how to normalize the data, defaults to standardize}

\item{train}{proportion of the data used for training}
}
\value{
model and test performance as a list
}
\description{
Train a k-nearest neighbors model on a comparison data frame to classify points that should be dropped
}
