% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{dev_add_to_gitignore}
\alias{dev_add_to_gitignore}
\title{Add big files to a .gitignore file}
\usage{
dev_add_to_gitignore(data_dir)
}
\arguments{
\item{data_dir}{directory of animal data files}
}
\description{
Add big files to a .gitignore file
}
