% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{dist_points_to_water}
\alias{dist_points_to_water}
\title{Find distance between n points and water objects}
\usage{
dist_points_to_water(points, xlim = c(0, 25), ylim = c(0, 25), water)
}
\arguments{
\item{points}{number of points to generate, or data.frame with lat/lon coordinates}

\item{xlim}{latitude bounds if points is a number}

\item{ylim}{longitude bounds if points is a number}

\item{water}{list of sf geoms representing water locations}
}
\value{
data.frame containing coordinates and their corresponding closest water sources
}
\description{
Find distance between n points and water objects
}
