% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{export_animal_elevation}
\alias{export_animal_elevation}
\title{Export modeled elevation data from existing animal data file}
\usage{
export_animal_elevation(rds_path, out_path)
}
\arguments{
\item{rds_path}{animal tracking data file to model elevation from}

\item{out_path}{exported file path}
}
\value{
list of data frames with gps data augmented by elevation
}
\description{
Export modeled elevation data from existing animal data file
}
