% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_elevation.R
\name{export_cow_elevation}
\alias{export_cow_elevation}
\title{Export modeled elevation data from existing cow data file}
\usage{
export_cow_elevation(rds_path, out_path)
}
\arguments{
\item{rds_path}{cow data file to model elevation from}

\item{out_path}{exported file path}
}
\description{
Export modeled elevation data from existing cow data file
}
