% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_elevation.R
\name{export_elevation}
\alias{export_elevation}
\title{Save high resolution elevation data to computer as .rds}
\usage{
export_elevation(out_path)
}
\arguments{
\item{out_path}{file path to save elevation data to}
}
\description{
Save high resolution elevation data to computer as .rds
}
