% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{ffimp}
\alias{ffimp}
\title{Feed-forward imputation on a comparison data frame
Helper function for detect_peak_knn}
\usage{
ffimp(df_comparison)
}
\arguments{
\item{df_comparison}{output of compare_flags}
}
\value{
imputed data frame
}
\description{
Feed-forward imputation on a comparison data frame
Helper function for detect_peak_knn
}
