% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_storage.R
\name{get_data_from_meta}
\alias{get_data_from_meta}
\title{Get animal data set from specified meta. 
If date range is invalid, automatically returns all animal data specified by meta_df.}
\usage{
get_data_from_meta(meta_df, min_date, max_date)
}
\arguments{
\item{meta_df}{data frame of specified meta}

\item{min_date}{minimum date specified by user}

\item{max_date}{maximum date specified by user}
}
\value{
df of animal data from specified meta
}
\description{
Get animal data set from specified meta. 
If date range is invalid, automatically returns all animal data specified by meta_df.
}
