% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{get_elevation}
\alias{get_elevation}
\title{Retrieve and save high resolution elevation data for the region of analysis from the internet}
\usage{
get_elevation(latmin, latmax, lonmin, lonmax, out_dir, zoom = 12,
  zone = 11)
}
\arguments{
\item{latmin}{minimum latitude for bounding box (degrees)}

\item{latmax}{maximum latitude for bounding box (degrees)}

\item{lonmin}{minimum longitude for bounding box (degrees)}

\item{lonmax}{maximum longitude for bounding box (degrees)}

\item{out_dir}{folder path to save the elevation data}

\item{zoom}{level of zoom, defaults to 12}

\item{zone}{geographic zone, defaults to 11}
}
\value{
elevation data as spatial points
}
\description{
Retrieve and save high resolution elevation data for the region of analysis from the internet
}
