% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_storage.R
\name{get_meta}
\alias{get_meta}
\title{Generate metadata for an animal data frame -
filename, site, date min/max, animals, min/max lat/longitude, storage location}
\usage{
get_meta(df, file_id, dtype, file_name, site, ani_id, storage_loc)
}
\arguments{
\item{df}{clean animal data frame}

\item{file_id}{ID number of source of animal data frame}

\item{dtype}{igotu or columbus}

\item{file_name}{.csv source of animal data frame}

\item{site}{physical source of animal data}

\item{ani_id}{ID of animal found in data frame}

\item{storage_loc}{.rds storage location of animal data frame}
}
\value{
df of metadata for animal data frame
}
\description{
Generate metadata for an animal data frame -
filename, site, date min/max, animals, min/max lat/longitude, storage location
}
