% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevation.R
\name{get_spdf}
\alias{get_spdf}
\title{Generate spatial points data frame from existing GPS data (.rds)}
\usage{
get_spdf(rds_path)
}
\arguments{
\item{rds_path}{path of GPS data (.rds)}
}
\value{
spatial points data frame
}
\description{
Generate spatial points data frame from existing GPS data (.rds)
}
