% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_elevation.R
\name{histogram_cow_elevation}
\alias{histogram_cow_elevation}
\title{Generate a histogram of the distribution of modeled elevation - measured altitude}
\usage{
histogram_cow_elevation(csv_path)
}
\arguments{
\item{csv_path}{path of csv GPS data to model elevation from}
}
\value{
histogram of the distribution of modeled elevation - measured altitude
}
\description{
Generate a histogram of the distribution of modeled elevation - measured altitude
}
