% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{histogram_time}
\alias{histogram_time}
\title{Generates a histogram to visualize the distribution of time
between GPS measurements.}
\usage{
histogram_time(rds_path)
}
\arguments{
\item{rds_path}{Path of .rds cow data file to read in}
}
\value{
distribution of time between GPS measurements, as a histogram
}
\description{
Generates a histogram to visualize the distribution of time
between GPS measurements.
}
\examples{
# Histogram of GPS measurement time differences for demo data .rds

histogram_time(system.file("extdata", "demo_nov19.rds", package = "animaltracker"))
}
