% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{histogram_time_unit}
\alias{histogram_time_unit}
\title{Generates a histogram to visualize the distribution of time between
GPS measurements by GPS unit.}
\usage{
histogram_time_unit(rds_path)
}
\arguments{
\item{rds_path}{Path of .rds animal data file to read in}
}
\value{
distribution of time between GPS measurements by GPS unit, as a histogram
}
\description{
Generates a histogram to visualize the distribution of time between
GPS measurements by GPS unit.
}
