% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{join_summaries}
\alias{join_summaries}
\title{Joins two animal data frame summaries by a column and appends differences}
\usage{
join_summaries(correct_summary, candidate_summary, by_str, daily = F)
}
\arguments{
\item{correct_summary}{summary df of reference dataset, returned by summarise_anidf}

\item{candidate_summary}{summary df of dataset to be compared to reference, returned by summarise_anidf}

\item{by_str}{column to join by as a string, null if daily=T}

\item{daily}{whether to group by both GPS and Date for daily summary, defaults to False}
}
\description{
Joins two animal data frame summaries by a column and appends differences
}
\examples{
# Join date summaries of unfiltered and filtered demo data
\donttest{
\dontrun{
## Get elevation 
elevfile <- system.file("extdata/elev", "USA_msk_alt.zip", package="animaltracker")
elev <- read_zip_to_rasters(elevfile )

## Get elevation data for unfiltered demo
unfiltered_elev <- lookup_elevation(elev, demo_unfiltered, zoom=11, 
get_slope=FALSE, get_aspect=FALSE)

## Get elevation data for filtered demo
filtered_elev <- lookup_elevation(elev, demo_filtered, zoom=11, get_slope=FALSE, get_aspect=FALSE)

## Summarise unfiltered demo by date
unfiltered_summary <- summarise_anidf(unfiltered_elev, Date, Latitude, Longitude, 
Distance, Course, Rate, Elevation, daily=FALSE)

## Summarise filtered demo by date
filtered_summary <- summarise_anidf(filtered_elev, Date, Latitude, Longitude, 
Distance, Course, Rate, Elevation, daily=FALSE)

## Join
join_summaries(unfiltered_summary, filtered_summary, "Date", daily=F)

}
}
}
