% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{kmz_to_sf}
\alias{kmz_to_sf}
\title{Convert kml coordinates to sf for mapping and calculations}
\usage{
kmz_to_sf(kmz_element, shift = c(0, 0))
}
\arguments{
\item{kmz_element}{kmz object containing coordinates}

\item{shift}{optional length 2 vector of coordinates to shift kml by}
}
\description{
Convert kml coordinates to sf for mapping and calculations
}
