% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{knnclassifier}
\alias{knnclassifier}
\title{k-nearest neighbors classifier for data validation}
\format{A Learner object}
\usage{
knnclassifier
}
\description{
k-nearest neighbors classifier for data validation
}
\keyword{datasets}
