% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{lookup_elevation_aws}
\alias{lookup_elevation_aws}
\title{Add elevation data from public AWS terrain tiles to long/lat coordinates of animal gps data}
\usage{
lookup_elevation_aws(anidf, zoom = 12)
}
\arguments{
\item{anidf}{animal tracking dataframe}

\item{zoom}{level of zoom, defaults to 12}
}
\value{
original data frame, with Elevation column appended
}
\description{
Add elevation data from public AWS terrain tiles to long/lat coordinates of animal gps data
}
\examples{
data(demo)
xelev <- lookup_elevation(demo, zoom = 12)
plot(xelev$Altitude, xelev$Elevation)
}
