% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{lookup_elevation_file}
\alias{lookup_elevation_file}
\title{Add elevation data from terrain tiles to long/lat coordinates of animal gps data}
\usage{
lookup_elevation_file(elev, anidf, zoom = 11, get_slope = TRUE,
  get_aspect = TRUE)
}
\arguments{
\item{elev}{elevation data as terrain tiles}

\item{anidf}{animal tracking dataframe}

\item{zoom}{level of zoom, defaults to 11}

\item{get_slope}{logical, whether to compute slope (in degrees), defaults to true}

\item{get_aspect}{logical, whether to compute aspect (in degrees), defaults to true}
}
\value{
original data frame, with terrain column(s) appended
}
\description{
Add elevation data from terrain tiles to long/lat coordinates of animal gps data
}
\examples{
# Add elevation data to demo data frame
\donttest{
\dontrun{
## Get elevation
elev <- read_zip_to_rasters(system.file("extdata", "elev/USA_msk_alt.zip", package="animaltracker"))

## Lookup with slope and aspect
lookup_elevation(elev, read.csv(system.file("extdata", 
"demo_aug19/Bannock_2017_101_1149.csv", package = "animaltracker"), skipNul = TRUE), 
zoom = 11, get_slope = TRUE, get_aspect = TRUE)
}
}
}
