% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{predict_peak_knn}
\alias{predict_peak_knn}
\title{Evaluate a k-nearest neighbors model on a comparison data frame to classify points that should be dropped}
\usage{
predict_peak_knn(df_comparison, model)
}
\arguments{
\item{df_comparison}{output of compare_flags}

\item{model}{output of train_peak_knn}
}
\value{
df_comparison with predictions
}
\description{
Evaluate a k-nearest neighbors model on a comparison data frame to classify points that should be dropped
}
