% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{process_elevation}
\alias{process_elevation}
\title{Export modeled elevation data from existing animal data file}
\usage{
process_elevation(rds_path, out_path, zoom = 12)
}
\arguments{
\item{rds_path}{animal tracking data file to model elevation from}

\item{out_path}{exported file path}

\item{zoom}{level of zoom, defaults to 12}
}
\value{
list of data frames with gps data augmented by elevation
}
\description{
Export modeled elevation data from existing animal data file
}
