% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{quantile_time}
\alias{quantile_time}
\title{Determines the GPS measurement time value difference values
roughly corresponding to quantiles with .05 intervals.}
\usage{
quantile_time(rds_path)
}
\arguments{
\item{rds_path}{Path of .rds animal data file to read in}
}
\value{
approximate time difference values corresponding to quantiles (.05 intervals)
}
\description{
Determines the GPS measurement time value difference values
roughly corresponding to quantiles with .05 intervals.
}
\examples{
# Read in .rds of demo data and calculate time difference quantiles

quantile_time(system.file("extdata", "demo_nov19.rds", package = "animaltracker"))
}
