% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs_module.R
\name{reactivePickerOutput}
\alias{reactivePickerOutput}
\title{Shiny Module UI output for the animaltracker app's dynamic dropdown selections.}
\usage{
reactivePickerOutput(id)
}
\arguments{
\item{id}{chosen ID of UI output}
}
\description{
Shiny Module UI output for the animaltracker app's dynamic dropdown selections.
}
