% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_module.R
\name{reactivePlot}
\alias{reactivePlot}
\title{'shiny' module server-side UI generator for the animaltracker app's summary statistics tables.}
\usage{
reactivePlot(input, output, session, plot_type, dat)
}
\arguments{
\item{input}{'shiny' server input, automatically populated}

\item{output}{'shiny' server output, automatically populated}

\item{session}{'shiny' server session, automatically populated}

\item{plot_type}{plot type to generate}

\item{dat}{animal data frame}
}
\value{
'shiny' renderPlot object
}
\description{
'shiny' module server-side UI generator for the animaltracker app's summary statistics tables.
}
