% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_module.R
\name{reactivePlotOutput}
\alias{reactivePlotOutput}
\title{'shiny' module UI output for the animaltracker app's plots tab.}
\usage{
reactivePlotOutput(id)
}
\arguments{
\item{id}{chosen ID of UI output}
}
\value{
'shiny' plotOutput object
}
\description{
'shiny' module UI output for the animaltracker app's plots tab.
}
