% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs_module.R
\name{reactiveRange}
\alias{reactiveRange}
\title{'shiny' module server-side UI generator for the animaltracker app's coordinate range input.}
\usage{
reactiveRange(input, output, session, type, dat)
}
\arguments{
\item{input}{'shiny' server input, automatically populated}

\item{output}{'shiny' server output, automatically populated}

\item{session}{'shiny' server session, automatically populated}

\item{type}{latitude or longitude}

\item{dat}{animal data frame}
}
\value{
'shiny' renderUI object for coordinate range input
}
\description{
'shiny' module server-side UI generator for the animaltracker app's coordinate range input.
}
