% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs_module.R
\name{reactiveRangeOutput}
\alias{reactiveRangeOutput}
\title{'shiny' module UI output for the animaltracker app's coordinate range input.}
\usage{
reactiveRangeOutput(id)
}
\arguments{
\item{id}{chosen ID of UI output}
}
\value{
'shiny' uiOutput for coordinate range input
}
\description{
'shiny' module UI output for the animaltracker app's coordinate range input.
}
