% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{read_gps}
\alias{read_gps}
\title{Reads a GPS dataset of unknown format at location filename}
\usage{
read_gps(filename)
}
\arguments{
\item{filename}{location of the GPS dataset}
}
\value{
list containing the dataset as a df and the format
}
\description{
Reads a GPS dataset of unknown format at location filename
}
