% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{read_zip_to_rasters}
\alias{read_zip_to_rasters}
\title{Read an archive of altitude mask files and convert the first file into a raster object}
\usage{
read_zip_to_rasters(filename, exdir = "inst/extdata/elev")
}
\arguments{
\item{filename}{path of altitude mask file archive}

\item{exdir}{path to extract files}
}
\value{
the first altitude mask file as a raster object
}
\description{
Read an archive of altitude mask files and convert the first file into a raster object
}
