% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_shiny_animaltracker}
\alias{run_shiny_animaltracker}
\title{Run the animaltracker 'shiny' app by calling this function.
Depending on the size of input files, it may be advisable to increase the maximum request size.}
\usage{
run_shiny_animaltracker(browser = TRUE, showcase = FALSE)
}
\arguments{
\item{browser}{logical, whether to launch the app in your default browser (defaults to TRUE)}

\item{showcase}{logical, whether to launch the app in 'showcase' mode (defaults to FALSE)}
}
\value{
None
}
\description{
Run the animaltracker 'shiny' app by calling this function.
Depending on the size of input files, it may be advisable to increase the maximum request size.
}
