% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_app.R
\name{run_validation_app}
\alias{run_validation_app}
\title{Run the 'shiny' validation app.
Depending on the size of input files, it may be advisable to increase the maximum request size.}
\usage{
run_validation_app()
}
\value{
None
}
\description{
Run the 'shiny' validation app.
Depending on the size of input files, it may be advisable to increase the maximum request size.
}
