% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_storage.R
\name{save_meta}
\alias{save_meta}
\title{Save metadata to a data frame and return it}
\usage{
save_meta(meta_df, file_meta)
}
\arguments{
\item{meta_df}{the data frame to store metadata in}

\item{file_meta}{meta for a .csv file generated by get_meta}
}
\description{
Save metadata to a data frame and return it
}
