% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs_module.R
\name{staticPicker}
\alias{staticPicker}
\title{Shiny Module server-side UI generator for the animaltracker app's basic dropdown selections.}
\usage{
staticPicker(
  input,
  output,
  session,
  selected_ani,
  text,
  choices,
  min_selected,
  max_selected
)
}
\arguments{
\item{input}{Shiny server input, automatically populated}

\item{output}{Shiny server output, automatically populated}

\item{session}{Shiny server session, automatically populated}

\item{selected_ani}{selected animals from animaltracker app input}

\item{text}{title for picker}

\item{choices}{vector of possible choices for picker}

\item{min_selected}{index of lowest selected value in choices}

\item{max_selected}{index of highest selected value in choices}
}
\description{
Shiny Module server-side UI generator for the animaltracker app's basic dropdown selections.
}
