% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs_module.R
\name{staticPickerOutput}
\alias{staticPickerOutput}
\title{Shiny Module UI output for the animaltracker app's basic dropdown selections.}
\usage{
staticPickerOutput(id)
}
\arguments{
\item{id}{chosen ID of UI output}
}
\value{
'shiny' uiOutput object for dropdown selection
}
\description{
Shiny Module UI output for the animaltracker app's basic dropdown selections.
}
