% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_module.R
\name{stats}
\alias{stats}
\title{'shiny' module server-side UI generator for the animaltracker app's summary statistics tables.}
\usage{
stats(
  input,
  output,
  session,
  selected_cols,
  selected_stats,
  col_name,
  col,
  dat
)
}
\arguments{
\item{input}{'shiny' server input, automatically populated}

\item{output}{'shiny' server output, automatically populated}

\item{session}{'shiny' server session, automatically populated}

\item{selected_cols}{selected columns from animaltracker app input}

\item{selected_stats}{selected summary statistics from animaltracker app input}

\item{col_name}{column name to compute summary statistics}

\item{col}{column to compute summary statistics}

\item{dat}{animal data frame containing col}
}
\value{
'shiny' renderTable object for table
}
\description{
'shiny' module server-side UI generator for the animaltracker app's summary statistics tables.
}
