% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_storage.R
\name{store_batch_list}
\alias{store_batch_list}
\title{Generates basic metadata about a directory of animal data files and stores the files as data frames as a list with the meta}
\usage{
store_batch_list(data_dir)
}
\arguments{
\item{data_dir}{location of animal data files, in list format}
}
\value{
a list of animal data frames with information about the data
}
\description{
Generates basic metadata about a directory of animal data files and stores the files as data frames as a list with the meta
}
