% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{summarise_anidf}
\alias{summarise_anidf}
\title{Calculates summary statistics for an animal data frame}
\usage{
summarise_anidf(anidf, by, lat, long, dist, course, rate, elev, daily = F)
}
\arguments{
\item{anidf}{the animal data frame}

\item{by}{column to group by, null if daily=T}

\item{lat}{latitude column}

\item{long}{longitude column}

\item{dist}{distance column}

\item{course}{course column}

\item{rate}{rate column}

\item{elev}{elevation column}

\item{daily}{whether to group by both GPS and Date for daily summary, defaults to False}
}
\description{
Calculates summary statistics for an animal data frame
}
\examples{
# Summary of demo data by date

summarise_anidf(demo, Date, Latitude, Longitude, Distance, Course, Rate, Elevation, daily=FALSE)

}
