% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{summarise_col}
\alias{summarise_col}
\title{Get summary statistics for a single column in an animal data frame}
\usage{
summarise_col(df, col)
}
\arguments{
\item{df}{animal data frame}

\item{col}{column to get summary stats for, as a string}
}
\value{
data frame of summary stats for col
}
\description{
Get summary statistics for a single column in an animal data frame
}
\examples{
# Get summary statistics for Distance column of demo data

summarise_col(demo, Distance)

}
