% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{summarise_unit}
\alias{summarise_unit}
\title{Summarise a number of animal datasets by GPS unit}
\usage{
summarise_unit(rds_path)
}
\arguments{
\item{rds_path}{Path of .rds cow data file to read in}
}
\value{
summary statistics for animals by GPS unit
}
\description{
Summarise a number of animal datasets by GPS unit
}
\examples{
# Read in .rds of demo data and summarise by GPS unit

summarise_unit(system.file("extdata", "demo_aug19.rds", package = "animaltracker"))
}
