% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{summarize_unit}
\alias{summarize_unit}
\title{Summarize by GPS unit}
\usage{
summarize_unit(rds_path)
}
\arguments{
\item{rds_path}{Path of .rds cow data file to read in}
}
\value{
summary statistics for animals by GPS unit
}
\description{
Summarize by GPS unit
}
