% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs_module.R
\name{time}
\alias{time}
\title{'shiny' module server-side UI generator for the animaltracker app's time input.}
\usage{
time(input, output, session, type, meta, selected_ani)
}
\arguments{
\item{input}{'shiny' server input, automatically populated}

\item{output}{'shiny' server output, automatically populated}

\item{session}{'shiny' server session, automatically populated}

\item{type}{min or max}

\item{meta}{animal metadata from app, must be non-empty for time input to display}

\item{selected_ani}{selected animals from app, must be non-empty for time to display}
}
\value{
'shiny' renderUI object for time input
}
\description{
'shiny' module server-side UI generator for the animaltracker app's time input.
}
