% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs_module.R
\name{timeOutput}
\alias{timeOutput}
\title{'shiny' module UI output for the animaltracker app's time input}
\usage{
timeOutput(id)
}
\arguments{
\item{id}{chosen ID of UI output}
}
\value{
'shiny' uiOutput for time input
}
\description{
'shiny' module UI output for the animaltracker app's time input
}
