% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{violin_compare}
\alias{violin_compare}
\title{Compares summary statistics from two datasets as side-by-side violin plots}
\usage{
violin_compare(df_summary, by, col_name, out)
}
\arguments{
\item{df_summary}{data frame of summary statistics from both datasets to be compared}

\item{by}{GPS or Date}

\item{col_name}{variable in df_summary to be used for the y-axis, as a string}

\item{out}{file name to save plot}
}
\value{
side-by-side violin plots
}
\description{
Compares summary statistics from two datasets as side-by-side violin plots
}
