% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{boxplot_altitude}
\alias{boxplot_altitude}
\title{Generates a boxplot to visualize the distribution of altitude
by GPS.}
\usage{
boxplot_altitude(rds_path)
}
\arguments{
\item{rds_path}{Path of .rds animal data file to read in}
}
\value{
overall boxplot of altitude by GPS
}
\description{
Generates a boxplot to visualize the distribution of altitude
by GPS.
}
