% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{clean_export_files}
\alias{clean_export_files}
\title{Cleans all animal GPS datasets (in .csv format) in a chosen directory. Optionally exports the clean data as spreadsheets, a single .rds data file, or as a list of data frames}
\usage{
clean_export_files(data_dir, cleaned_filename = "data/animal_data.rds",
  cleaned_dir = "data/processed")
}
\arguments{
\item{data_dir}{directory of GPS tracking files (in csv)}

\item{out_path}{full name of output file (ending in .rds)}

\item{processed_dir}{optional directory to save the processed GPS datasets as spreadsheets (.csv)}

\item{saveRDS}{logical, whether to save the processed data to a single R object (.rds)}

\item{saveCSV}{logical, whether to save the processed data to spreadsheets in}
}
\description{
Cleans all animal GPS datasets (in .csv format) in a chosen directory. Optionally exports the clean data as spreadsheets, a single .rds data file, or as a list of data frames
}
