% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{clean_export_files}
\alias{clean_export_files}
\title{Cleans all cow datasets in a chosen directory and exports them as a single .rds file}
\usage{
clean_export_files(data_dir, out_path)
}
\arguments{
\item{data_dir}{directory of cow data files}

\item{out_path}{name of output file, must end in .rds}
}
\description{
Cleans all cow datasets in a chosen directory and exports them as a single .rds file
}
