% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{clean_location_data}
\alias{clean_location_data}
\title{Cleans a raw animal GPS dataset, implementing a standardized procedure to remove impossible values}
\usage{
clean_location_data(df, aniid = NA, gpsid = NA, maxrate = 84,
  maxcourse = 100, maxdist = 840, maxtime = 100, timezone = "UTC")
}
\arguments{
\item{df}{data frame in standardized format (e.g., from a raw spreadsheet)}

\item{aniid}{identification code for the animal}

\item{gpsid}{identification code for the GPS device}

\item{maxrate}{maximum rate of travel (meters/minute) between consecutive points}

\item{maxcourse}{maximum distance (meters) between consecutive points}

\item{maxdist}{maximum geographic distance (meters) between consecutive points}

\item{maxtime}{maximum time (minutes) between consecutive points}

\item{timezone}{identification code for the animal}
}
\description{
Cleans a raw animal GPS dataset, implementing a standardized procedure to remove impossible values
}
