% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_storage.R
\name{clean_store_batch}
\alias{clean_store_batch}
\title{Cleans a directory of animal data files and stores them locally in rds format}
\usage{
clean_store_batch(data_info, autocleans, filters, get_slope, get_aspect,
  min_lat, max_lat, min_long, max_long)
}
\arguments{
\item{data_info}{list of animal data frames with information about the data, generated by store_batch}

\item{autocleans}{automatically clean data with ts_clean, defaults to true}

\item{filters}{filter bad data points, defaults to true}

\item{get_slope}{logical, whether to compute slope (in degrees)}

\item{get_aspect}{logical, whether to compute aspect (in degrees)}

\item{min_lat}{minimum latitude for filtering}

\item{max_lat}{maximum latitude for filtering}

\item{min_long}{minimum longitude for filtering}

\item{max_long}{maximum longitude for filtering}
}
\value{
df of metadata for animal file directory
}
\description{
Cleans a directory of animal data files and stores them locally in rds format
}
