% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{compare_flags}
\alias{compare_flags}
\title{Joins and reformats two animal data frames for the purpose of flag comparison}
\usage{
compare_flags(correct, candidate)
}
\arguments{
\item{correct}{reference df}

\item{candidate}{df to be compared to the reference}
}
\description{
Joins and reformats two animal data frames for the purpose of flag comparison
}
