% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{compare_flags}
\alias{compare_flags}
\title{Joins and reformats two animal data frames for the purpose of flag comparison}
\usage{
compare_flags(correct, candidate)
}
\arguments{
\item{correct}{reference df}

\item{candidate}{df to be compared to the reference}
}
\value{
joined and reformatted df
}
\description{
Joins and reformats two animal data frames for the purpose of flag comparison
}
\examples{
\donttest{
\dontrun{
# Join and reformat unfiltered demo data and filtered demo data

## Get elevation data for unfiltered demo
unfiltered_elev <- lookup_elevation_aws(demo_unfiltered, zoom=1, 
get_slope=FALSE, get_aspect=FALSE)

## Get elevation data for filtered demo
filtered_elev <- lookup_elevation_aws(demo_filtered, zoom=1, get_slope=FALSE, get_aspect=FALSE)

compare_flags(unfiltered_elev, filtered_elev)
}
}
}
