% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{compare_summarise_daily}
\alias{compare_summarise_daily}
\title{Compares two animal datasets and calculates daily summary statistics. 
GPS, date, lat, long, course, distance, rate, elevation column names should match.}
\usage{
compare_summarise_daily(correct, candidate, out)
}
\arguments{
\item{correct}{reference df}

\item{candidate}{df to be compared to the reference}

\item{out}{desired file name of .csv output summary}
}
\value{
summary df
}
\description{
Compares two animal datasets and calculates daily summary statistics. 
GPS, date, lat, long, course, distance, rate, elevation column names should match.
}
