% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{compare_summarise_data}
\alias{compare_summarise_data}
\title{Compares two animal datasets and calculates summary statistics. 
GPS, date, lat, long, course, distance, rate, elevation column names should match.}
\usage{
compare_summarise_data(correct, candidate, gps_out, date_out)
}
\arguments{
\item{correct}{reference df}

\item{candidate}{df to be compared to the reference}

\item{gps_out}{desired file name of .csv output summary by GPS collar}

\item{date_out}{desired file name of .csv output summary by date}
}
\value{
list containing gps_out and date_out as dfs
}
\description{
Compares two animal datasets and calculates summary statistics. 
GPS, date, lat, long, course, distance, rate, elevation column names should match.
}
