% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{detect_peak_modz}
\alias{detect_peak_modz}
\title{Alternative implementation of the robust peak detection algorithm by van Brakel 2014 
Classifies data points with modified z-scores greater than max_score as outliers ccording to Iglewicz and Hoaglin 1993}
\usage{
detect_peak_modz(df_comparison, lag = 5, max_score = 3.5)
}
\arguments{
\item{df_comparison}{output of compare_flags}

\item{lag}{width of interval to compute rolling median and MAD, defaults to 5}

\item{max_score}{modified z-score cutoff to classify observations as outliers, defaults to 3.5}
}
\value{
df with classifications
}
\description{
Alternative implementation of the robust peak detection algorithm by van Brakel 2014 
Classifies data points with modified z-scores greater than max_score as outliers ccording to Iglewicz and Hoaglin 1993
}
\examples{
\donttest{
\dontrun{
# Join and reformat unfiltered demo data and filtered demo data

## Get elevation data for unfiltered demo
unfiltered_elev <- lookup_elevation_aws(demo_unfiltered, zoom=1, 
get_slope=FALSE, get_aspect=FALSE)

## Get elevation data for filtered demo
filtered_elev <- lookup_elevation_aws(demo_filtered, zoom=1, get_slope=FALSE, get_aspect=FALSE)

## Get comparison df
comparison <- compare_flags(unfiltered_elev, filtered_elev)

detect_peak_modz(comparison, lag = 5, max_score = 3.5)
}
}
}
