% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{detect_peak_modz}
\alias{detect_peak_modz}
\title{Alternative implementation of the robust peak detection algorithm by van Brakel 2014 
Classifies data points with modified z-scores greater than 3.5 as outliers ccording to Iglewicz and Hoaglin 1993}
\usage{
detect_peak_modz(df_comparison, lag = 5)
}
\arguments{
\item{df_comparison}{output of compare_flags}

\item{lag}{width of interval to compute rolling median and MAD}
}
\description{
Alternative implementation of the robust peak detection algorithm by van Brakel 2014 
Classifies data points with modified z-scores greater than 3.5 as outliers ccording to Iglewicz and Hoaglin 1993
}
