% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{join_summaries}
\alias{join_summaries}
\title{Helper function for compare_summarise_data
Joins two animal dataset summaries by a column and appends differences}
\usage{
join_summaries(correct_summary, candidate_summary, by, daily = F)
}
\arguments{
\item{correct_summary}{summary df of reference dataset, returned by summarise_anidf}

\item{candidate_summary}{summary df of dataset to be compared to reference, returned by summarise_anidf}

\item{by}{column to join by , null if daily=T}

\item{daily}{whether to group by both GPS and Date for daily summary, defaults to False}
}
\description{
Helper function for compare_summarise_data
Joins two animal dataset summaries by a column and appends differences
}
