% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{lookup_elevation_aws}
\alias{lookup_elevation_aws}
\title{Add elevation data from public AWS terrain tiles to long/lat coordinates of animal gps data}
\usage{
lookup_elevation_aws(anidf, zoom = 12, get_slope = TRUE,
  get_aspect = TRUE)
}
\arguments{
\item{anidf}{animal tracking dataframe}

\item{zoom}{level of zoom, defaults to 12}

\item{get_slope}{logical, whether to compute slope (in degrees), defaults to true}

\item{get_aspect}{logical, whether to compute aspect (in degrees), defaults to true}
}
\value{
original data frame, with Elevation column appended
}
\description{
Add elevation data from public AWS terrain tiles to long/lat coordinates of animal gps data
}
