% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{lookup_elevation_file}
\alias{lookup_elevation_file}
\title{Add elevation data from terrain tiles to long/lat coordinates of animal gps data}
\usage{
lookup_elevation_file(elev, anidf, zoom = 11, get_slope = TRUE,
  get_aspect = TRUE)
}
\arguments{
\item{elev}{elevation data as raster}

\item{anidf}{animal tracking dataframe}

\item{zoom}{level of zoom, defaults to 11}

\item{get_slope}{logical, whether to compute slope (in degrees), defaults to true}

\item{get_aspect}{logical, whether to compute aspect (in degrees), defaults to true}
}
\value{
original data frame, with terrain column(s) appended
}
\description{
Add elevation data from terrain tiles to long/lat coordinates of animal gps data
}
