% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{model_animal_elevation}
\alias{model_animal_elevation}
\title{Model elevation from GPS data (provided csv)}
\usage{
model_animal_elevation(csv_path)
}
\arguments{
\item{csv_path}{path of csv GPS data}
}
\value{
modeled elevation data
}
\description{
Model elevation from GPS data (provided csv)
}
