% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{process_elevation}
\alias{process_elevation}
\title{Export modeled elevation data from existing animal data file}
\usage{
process_elevation(
  zoom = 11,
  get_slope = TRUE,
  get_aspect = TRUE,
  in_path,
  out_path
)
}
\arguments{
\item{zoom}{level of zoom, defaults to 11}

\item{get_slope}{logical, whether to compute slope (in degrees), defaults to true}

\item{get_aspect}{logical, whether to compute aspect (in degrees), defaults to true}

\item{in_path}{animal tracking data file to model elevation from}

\item{out_path}{exported file path, .rds}
}
\value{
list of data frames with gps data augmented by elevation
}
\description{
Export modeled elevation data from existing animal data file
}
\examples{
# Export elevation data from demo .rds datasets
\donttest{
\dontrun{
process_elevation(zoom = 11, get_slope = TRUE, get_aspect = TRUE, 
in_path = system.file("extdata", "demo_nov19.rds", 
package = "animaltracker"), out_path = "demo_nov19_elev.rds")

}
}
}
