% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs_module.R
\name{reactivePicker}
\alias{reactivePicker}
\title{Shiny Module server-side UI generator for the animaltracker app's dynamic dropdown selections.}
\usage{
reactivePicker(
  input,
  output,
  session,
  type,
  req_list,
  text,
  min_selected = NULL,
  max_selected = NULL,
  multiple,
  options = NULL
)
}
\arguments{
\item{input}{Shiny server input, automatically populated}

\item{output}{Shiny server output, automatically populated}

\item{session}{Shiny server session, automatically populated}

\item{type}{purpose of picker - currently supported types are "site", "ani", and "recent"}

\item{req_list}{list of reactive statements required to display picker}

\item{text}{title for picker}

\item{min_selected}{index of lowest selected value in possible choices, should be null if type is "recent"}

\item{max_selected}{index of highest selected value in possible choices should be null if type is "recent"}

\item{multiple}{logical, whether to allow selecting multiple values}

\item{options}{options for shinyWidgets pickerInput}
}
\description{
Shiny Module server-side UI generator for the animaltracker app's dynamic dropdown selections.
}
