% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs_module.R
\name{reactivePickerOutput}
\alias{reactivePickerOutput}
\title{'shiny' module UI output for the animaltracker app's dynamic dropdown selections.}
\usage{
reactivePickerOutput(id)
}
\arguments{
\item{id}{chosen ID of UI output}
}
\value{
'shiny' uiOutput object for dropdown selection
}
\description{
'shiny' module UI output for the animaltracker app's dynamic dropdown selections.
}
