% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_shiny_animaltracker}
\alias{run_shiny_animaltracker}
\title{You can run the animaltracker Shiny app by calling this function.}
\usage{
run_shiny_animaltracker()
}
\description{
You can run the animaltracker Shiny app by calling this function.
}
