% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_shiny_animaltracker}
\alias{run_shiny_animaltracker}
\title{You can run the animaltracker Shiny app by calling this function.}
\usage{
run_shiny_animaltracker(rds_path)
}
\arguments{
\item{rds_path}{Path of Animal data file to input}
}
\description{
You can run the animaltracker Shiny app by calling this function.
}
