% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_module.R
\name{statsLabel}
\alias{statsLabel}
\title{'shiny' module server-side UI generator for the animaltracker app's summary statistics labels.}
\usage{
statsLabel(
  input,
  output,
  session,
  selected_cols,
  selected_stats,
  col_name,
  text
)
}
\arguments{
\item{input}{'shiny' server input, automatically populated}

\item{output}{'shiny' server output, automatically populated}

\item{session}{'shiny' server session, automatically populated}

\item{selected_cols}{selected columns from animaltracker app input}

\item{selected_stats}{selected summary statistics from animaltracker app input}

\item{col_name}{column name to compute summary statistics}

\item{text}{text of summary statistics label}
}
\value{
'shiny' renderUI object for label
}
\description{
'shiny' module server-side UI generator for the animaltracker app's summary statistics labels.
}
