% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_module.R
\name{statsLabelOutput}
\alias{statsLabelOutput}
\title{'shiny' Module UI output for the animaltracker app's summary statistics labels.}
\usage{
statsLabelOutput(id)
}
\arguments{
\item{id}{chosen ID of UI output}
}
\value{
'shiny' uiOutput object for label
}
\description{
'shiny' Module UI output for the animaltracker app's summary statistics labels.
}
