% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_module.R
\name{statsOutput}
\alias{statsOutput}
\title{Shiny Module UI output for the animaltracker app's summary statistics tables.}
\usage{
statsOutput(id)
}
\arguments{
\item{id}{chosen ID of UI output}
}
\description{
Shiny Module UI output for the animaltracker app's summary statistics tables.
}
