% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{summarise_anidf}
\alias{summarise_anidf}
\title{Calculates summary statistics for an animal data frame}
\usage{
summarise_anidf(
  anidf,
  by,
  lat,
  long,
  dist,
  course,
  rate,
  elev = NULL,
  use_elev = TRUE,
  daily = FALSE
)
}
\arguments{
\item{anidf}{the animal data frame}

\item{by}{column to group by, null if daily=TRUE}

\item{lat}{latitude column}

\item{long}{longitude column}

\item{dist}{distance column}

\item{course}{course column}

\item{rate}{rate column}

\item{elev}{elevation column, must be defined when use_elev is true, otherwise NULL}

\item{use_elev}{logical, whether to include elevation in summary, defaults to true}

\item{daily}{whether to group by both GPS and Date for daily summary, defaults to false}
}
\value{
data frame of summary statistics for the animal data frame
}
\description{
Calculates summary statistics for an animal data frame
}
\examples{
# Summary of demo data by date

summarise_anidf(demo, Date, Latitude, Longitude, Distance, Course, Rate, Elevation, daily=FALSE)

}
