% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{summarise_anidf}
\alias{summarise_anidf}
\title{Helper function for compare_summarise_data and compare_summarise_daily
Calculates summary statistics for an animal dataset}
\usage{
summarise_anidf(anidf, by, lat, long, dist, course, rate, elev,
  daily = F)
}
\arguments{
\item{anidf}{the animal dataset}

\item{by}{column to group by, null if daily=T}

\item{lat}{latitude column}

\item{long}{longitude column}

\item{dist}{distance column}

\item{course}{course column}

\item{rate}{rate column}

\item{elev}{elevation column}

\item{daily}{whether to group by both GPS and Date for daily summary, defaults to False}
}
\description{
Helper function for compare_summarise_data and compare_summarise_daily
Calculates summary statistics for an animal dataset
}
