% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{violin_compare}
\alias{violin_compare}
\title{Compares summary statistics from two datasets as side-by-side violin plots}
\usage{
violin_compare(df_summary, by, col_name, out)
}
\arguments{
\item{df_summary}{data frame of summary statistics from both datasets to be compared}

\item{by}{GPS or Date}

\item{col_name}{variable in df_summary to be used for the y-axis, as a string}

\item{out}{file name to save plot}
}
\value{
side-by-side violin plots
}
\description{
Compares summary statistics from two datasets as side-by-side violin plots
}
\examples{
# Violin plot comparing unfiltered and filtered demo data summaries by date for a single variable
\donttest{
\dontrun{
## Get elevation
elevfile <- system.file("extdata/elev", "USA_msk_alt.zip", package="animaltracker")
elev <- read_zip_to_rasters(elevfile)

## Get elevation data for unfiltered demo
unfiltered_elev <- lookup_elevation(elev, demo_unfiltered, zoom=11, 
get_slope=FALSE, get_aspect=FALSE)

## Get elevation data for filtered demo
filtered_elev <- lookup_elevation(elev, demo_filtered, zoom=11, get_slope=FALSE, get_aspect=FALSE)

## Summarise unfiltered demo
unfiltered_summary <- summarise_anidf(unfiltered_elev, Date, Latitude, Longitude, 
Distance, Course, Rate, Elevation, daily=FALSE)

## Summarise filtered demo
filtered_summary <- summarise_anidf(filtered_elev, Date, Latitude, Longitude, 
Distance, Course, Rate, Elevation, daily=FALSE)

## Join
summary <- join_summaries(unfiltered_summary, filtered_summary, "Date", daily=FALSE)

## Violin plot

violin_compare(summary, Date, "meanElev", "ex_elev_violin.png")

}
}
}
